♻refactor: merged all lib files in /lib
