+++ src/app/api/user/route.ts
@@ -1,5 +1,5 @@
 import { NextResponse, type NextRequest } from 'next/server'
-import { prisma } from '@/libs/prisma'
+import { prisma } from '@/lib/prisma'
 
 export async function GET(req: NextRequest) {
   if (req) {
